Refactor into multiple address subcategories
