+++ src/aims_ui/page_controllers/b_multiple_matches/large_multiple_match/page_multiple_address_large_submit.py
@@ -12,7 +12,7 @@ from aims_ui.page_helpers.pages_location_utils import get_page_location
 from aims_ui.page_helpers.security_utils import check_user_has_access_to_page
 
 from .utils.multiple_match_api_utils import count_active_jobs
-from .utils.multiple_match_file_upload_utils import check_valid_upload, validate_job_name, validate_limit_parameter
+from ..utils.multiple_match_file_upload_utils import check_valid_upload, validate_job_name, validate_limit_parameter
 from .utils.submit_multiple_match_api import multiple_address_match
 
 page_name = 'multiple_address'
@@ -22,7 +22,7 @@ max_jobs = app.config.get('BM_MAX_JOBS')
 
 @login_required
 @app.route(f'/{page_name}', methods=['GET', 'POST'])
-def multiple_address():
+def multiple_address_large_submit():
   endpoints = get_endpoints(called_from=page_name)
   access = check_user_has_access_to_page(page_name)
   if access != True:
