+++ src/aims_ui/page_controllers/d_misc_functionality/download_handlers/large_multiple_address_download.py
@@ -9,7 +9,7 @@ from flask_login import login_required
 
 from aims_ui import app
 from aims_ui.app_helpers.classification_utils.autosuggest import get_classification_autosuggest_list, get_full_classification_autosuggest, get_classification_autosuggest_list_reversed
-from aims_ui.page_controllers.b_multiple_matches.utils.multiple_match_api_utils import job_url_if_authorised
+from aims_ui.page_controllers.b_multiple_matches.large_multiple_match.utils.multiple_match_api_utils import job_url_if_authorised
 from aims_ui.page_controllers.f_error_pages.page_error import page_error
 
 # For the gz download
