+++ src/aims_ui/page_controllers/d_misc_functionality/download_handlers/static_files_download.py
@@ -8,7 +8,7 @@ from flask import send_file
 from flask_login import login_required
 
 from aims_ui import app
-from aims_ui.page_controllers.b_multiple_matches.utils.multiple_match_api_utils import job_url_if_authorised
+from aims_ui.page_controllers.b_multiple_matches.large_multiple_match.utils.multiple_match_api_utils import job_url_if_authorised
 from aims_ui.page_controllers.f_error_pages.page_error import page_error
 
 # For the gz download
