+++ src/aims_ui/templates/aims_ui_pages/b_multiple_matches/small_multiple_match/multiple_address_small_submit.html
@@ -18,7 +18,7 @@
 <h1>Match Multiple Addresses</h1>
 
 {{ customMultipleMatchNav ({
-	'current_path': 'multiple_address_original',
+	'current_path': 'multiple_address_small_submit',
 	'bulk_limits': bulk_limits,
 	}) }}
 
@@ -89,7 +89,7 @@
 <!-- Single entry point for all page scripts. Add scripts to the main location below -->
 {{ runScriptWithPageName({
 		"page_name": page_name,
-		"script_location": '/static/js/b_multiple_matches/multiple_address_original/init.mjs',
+		"script_location": '/static/js/b_multiple_matches/multiple_address_small_submit/init.mjs',
 }) }}
