+++ src/aims_ui/templates/custom_multiple_match_nav/_macro.njk
@@ -12,7 +12,7 @@
     "itemsList": [
       {
         "title": "Under " + (params.bulk_limits.limit_mini_bulk|string) + " Addresses",
-        "url": 'multiple_address_original',
+        "url": 'multiple_address_small_submit',
       }
     ] + ([] if reduced else [
       {
